website: rename the "prefix" variable to "oldPrefix"
I realized that the choice of the previous prefix variable wasn't great.
Since it's the prefix that factors into public resource names, it should
contain the bits that will uniquify our various resources, as in the
"wwtdev" value that was suggested in the comments. We used "wwtcoreapp"
to create the current resources, though, which is weak since (for
instance) it doesn't distinguish dev/prod.

Renaming resources in azure is a big pain but in an effort to improve
things, we'll rename the old prefix and try to use better naming going
forward.