prod/constellations-cosmosdb.tf: sync up with db restoration
So, after creating this CosmosDB instance, I deleted it and created one
in the Azure Portal by restoring a backup of the development database.
The "new" database appears to be a bit weird, but it also does seem to
be populated with the content put together in the dev site. It might be
necessary that a "database migration" would have been a better way to do
this, but we'll see how well things work as is.